pyproject.toml
[project]
name = "changefacialexpression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "pillow>=11.1.0",
    "pip>=25.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.42.1",
]
